[ignore] Move module_lookup outside of /symbolication
