+++ crates/symbolicator-service/src/services/module_lookup.rs
@@ -5,7 +5,7 @@ use futures::future;
 use sentry::{Hub, SentryFutureExt};
 
 use symbolic::debuginfo::ObjectDebugSession;
-use symbolicator_sources::{FileType, ObjectType, SourceConfig};
+use symbolicator_sources::{FileType, ObjectId, ObjectType, SourceConfig};
 
 use crate::caching::{CacheEntry, CacheError};
 use crate::services::derived::DerivedCache;
@@ -16,11 +16,26 @@ use crate::services::ppdb_caches::{
 use crate::services::symcaches::{FetchSymCache, OwnedSymCache, SymCacheActor};
 use crate::types::{
     AllObjectCandidates, CompleteObjectInfo, CompleteStacktrace, ObjectFeatures, ObjectFileStatus,
-    RawFrame, RawStacktrace, Scope,
+    RawFrame, RawObjectInfo, RawStacktrace, Scope,
 };
 use crate::utils::addr::AddrMode;
 
-use super::object_id_from_object_info;
+fn object_id_from_object_info(object_info: &RawObjectInfo) -> ObjectId {
+    ObjectId {
+        debug_id: match object_info.debug_id.as_deref() {
+            None | Some("") => None,
+            Some(string) => string.parse().ok(),
+        },
+        code_id: match object_info.code_id.as_deref() {
+            None | Some("") => None,
+            Some(string) => string.parse().ok(),
+        },
+        debug_file: object_info.debug_file.clone(),
+        code_file: object_info.code_file.clone(),
+        debug_checksum: object_info.debug_checksum.clone(),
+        object_type: object_info.ty,
+    }
+}
 
 pub fn object_file_status_from_cache_entry<T>(cache_entry: &CacheEntry<T>) -> ObjectFileStatus {
     match cache_entry {
